+++ packages/terra-popup/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added requestFocus and releaseFocus props to support requesting and releasing focus.
 
 1.0.0 - (July 26, 2017)
 ------------------